Ensure last machine removed does not wait for hook
When the last machine is in a deleting state, this means that cluster is
removed also. In such scenario, waiting for draining is not feasible,
because it is performes only when node deletion is allowed. Which is
not, due to cluster removal. Cluster API prevents draining with the
"cluster is being deleted" error.

Signed-off-by: Danil-Grigorev <danil.grigorev@suse.com>