drop semanage call on nonexisting service
the rke2 install script installs the service in /etc/systemd/service
and the restorecon call already sets the proper context on it (but since
we start the service beforehand, that doesn't seem to be necessary..).

semanage(1) isn't always preinstalled in the operating system image, and
if this service exists, the relabeling at boot will restore the context.

Signed-off-by: Dirk Müller <dmueller@suse.com>